LayoutTests/platform/win/imported/w3c/web-platform-tests/cors/request-headers-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=x-print
CONSOLE MESSAGE: line 22: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=x-print,
CONSOLE MESSAGE: line 36: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=x-print, due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=x-print
CONSOLE MESSAGE: line 51: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=,y-lol,x-PriNT,%20,,,Y-PRINT
CONSOLE MESSAGE: line 59: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=,y-lol,x-PriNT,%20,,,Y-PRINT due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=,y-lol,x-PriNT,%20,,,Y-PRINT
Request headers


FAIL basic request header  A network error occurred.
FAIL Simple request headers need not be in allow-headers  A network error occurred.
PASS Unspecified request headers are disallowed 
FAIL Strange allowheaders (case insensitive)  A network error occurred.
PASS INVALID_STATE_ERR on setRequestHeader before open() 
FAIL INVALID_STATE_ERR on setRequestHeader after send()  A network error occurred.
